refactor(network): move files
